feat: add warning if number of uploads differ
Introduces a message when the number of uploads differ and there is a significant coverage drop.
The idea is that we can inform users that possibly some upload is missing, and that is the cause of the coverage drop.

I had to re-work part of the logic that counts the upload diff because the sessions in a report is a `dict` object, not a `list`. Thankfully we have many tests :E