chore: rewire experiment
After the Sentry metrics disappointed in producing quantiles we are switching to prometheus.
The query language is hard, but we know that it can be done.
I basically copied the code from `tasks.base` and modified it.

I also re-wired the experiment so that we can have more feature flags in there (if wanted)