fix: rewrite raw uploads empty and expired
The re-write on readable format is currently not working for expired or empty reports.
This is because we don't returned the raw report (that despite the name does have a little
but of processing to it) in the case that the report is empty or expired.
By passing the raw_parsed_report we should start rewriting them in the readable format.

This didn't fall in the "rewrite in error" cathegory because these scenarios are not
treated as errors.