fix+feat: update diff number of uploads comment
These changes affect the PR Comment "there's a different number of uploads
between BASE and HEAD" section. And include 2 changes:

1. A padding line in the template of the collapsible table so it will be
properly rendered in GitHub
2. A change in the logic to emit the comment, as discussed below.

When testing the feature manually I had the situation where an upload with a flag
was not sent and this reduced by a lot coverage. The issue is that there's only 1
upload with this flag, so the condition that _both_ BASE and HEAD have an upload was
not met.
Stil this is precisely the situation we want to cover.

To address this scenario then we are differentiating now flags that are carryforward
from regular flags. For Carryforward it is still the case that _both BASE and HEAD need
an upload for the difference to be considered.

For regular flags it suffices that BASE has a upload, and a larger number of uploads than
the HEAD (which was the case in my scenario).

Sharp coverage drop is stil a requirement, to reduce false positives.