fix: coverage_drop condition
THe `coverage_drop` comment condition looks for the project status threshold,
but the config might be a simple bool. This causes the notifier to fail with
and exception.

These changes fix that and add some more testing to cover this case (that should
have been added earlier, I know. My bad)