feat: update bundle_analysis comment requirerement
Updates the comment requirements for bundle_analysis.
Respects the `comment.bundle_change_threshold` defined in the YAML
and only notifies if changes are bigger than the threshold.
(threshold is expected to be an int. So technically [this PR](https://github.com/codecov/shared/pull/271) is a dependency of the present changes, because it does the
coercion into an integer, as expected)

Also adds the `"bundle_increase"` as a possible requirements. In this case
on top of the threshold the change also needs to cause the bundle to _increase_
in size for a comment to be sent.

closes: https://github.com/codecov/engineering-team/issues/2021