feat: don't select suspended apps
Currently we ignore if an app is suspended or not.
An app is suspended by the user.
If we select a suspended app we will fail to get an access_token for it
and eventually fail with `InstallationError`

These changes filter out suspended apps from being selected.
In case all apps for the user are suspended we skip notifications.
Other tasks would fail with `NoConfiguredAppsAvailable`.

closes https://github.com/codecov/internal-issues/issues/519