fix[process flakes]: filter fetched test instances
We were previously getting all test instances related to a given repo,
commit, branch combination. This is more than we actually need, we only
need the test instances that had a failed outcome OR have a test id
that matches an existing flake in our database. The goal of this change
is to filter out a bunch of test instances that passed that aren't
relevant to processing flakes.