chore: don't delete comments.
Currently the `require_changes` will deelte an existing comment if the requirements are not met but a commit exists. The commit vanishes without trace, which is not great UX.

To improve UX we are changing this behavior to bypass required changes if a comment already exists for that PR. We will update that comment.

We did verify that this does not trigger a new email being sent to users, so it should be fine.

closes: https://github.com/codecov/engineering-team/issues/2023