feat: use merge commit sha for was_squash
We want to use the merge commit sha provided by the repo providers APIs
to deduce whether the merge was done via regular merge or squash merge.

We can do this by getting the merge commit sha, then making an extra
API call to get it's parents. If it has multiple parents then it's a
regular merge, else it's a squash.

This commit introduces this as an experiment that is meant to run in
the background. Metrics and logs were introduced in order to observe the
correctness of the experiment.