feat: add BaseNotifier and NotifierResult
The BaseNotifier class will be used by the TestResultsNotifier and the
ErrorCommentNotifier. It provides a basic structure and utility for
sending the comment to Github.

The build_message method is not implemented and actual concrete
notifier classes are meant to extend this class and implement the
build_message method to generate the contents of the comment that will
be sent to Github when notify() is called.