feat: test results notifier extends base notifier
This commit changes the TestResultsNotifier to extend the BaseNotifier
class from helpers. A change that was made as a side effect is that
the BaseNotifier does not accept any arguments in the build_message or
notify methods so the class constructor was extended to accept the
payload that is used in the build_message method.