WIP: define the interface to create and send messages
Defines the `MessageStrategy` and use it in the notify service.
Notice that there was a circular dependency between the `MessageStrategy` and the
`NotificationContext` if the strategy was inside the context, so I encapsulated it one level above.

So now, inf the notify service, you specify the message context and its corresponding strategy per notification type. I think it's nice because we can technically combine different strategies with the same context (assuming there are more strategies in the future)