address review comments
Create a descriptor for notification context fields so that we fail with a
custom exception if that field was not loaded but we are trying to access it.

With the added benefit of reducing the amount of code a little bit :E

I added typehints to the fields (e.g. `pull: EnrichedPull = NotificationContextField[EnrichedPull]()`) because my VSCode langauge server was not being helpful with the generics... but leaving the generics in any case.