Gracefully handle nr=0 in JaCoCo report
JaCoCo currently sometimes output line 0 for coverage report, which is a confusing behaviour.
This occurs primarily when the Kotlin compiler cannot properly resolve synthetic functions.
Although this should ultimately be fixed in the Kotlin-JaCoCo report generation,
we're just going to skip the line to allow for the rest of the report processing to work.