chore: user config in notifications
These changes hide away the current_yaml from the Contexts.
The original idea was to not expose it, but I didn't do that detail correctly the 1st time.
Not it's hidden behind the `user_config` field. Builders still have full access.

The `user_config` field parses the config that will be used for commit statuses,
and also the config already used for the PR comment required changes.

There are 2 similar but different configurations:
* `required_changes_threshold` is the threshold to send the comment;
* `warning_threshold` is the threshold to fail status checks (emit warning by default)