chore: change `ComparisonLoader` init
The `ComparisonLoader` depended on a `EnrichedPull` object to be created.
This made sense in the PR Comment context where there's always a pull.
For the commit status we might not have a Pull, in which case we will default to
comparing to the current commit's parent.

This is still a comparison nonetheless, so most of the loader is the same.
I'm just changing how it is created and adding a `from_EnrichedPull` classmethod.

This is to be used in the bundle_analysis commit status