Simplify and fix debouncing of `Upload` task
This moves the logic to debounce multiple `Upload` tasks based `upload_processing_delay` to its own function.
It also fixes the debouncing logic so that it properly takes into account the previous task time.

Previously, assuming a 10-second delay, a previous upload 9 seconds in the past would delay the next upload by 10 seconds, for a total delta of 19. This now schedules the task in 1 second, so the total delay is at 10 seconds.