feat: commit_status context (for BA)
Creates the commit status context for bundle analysis.
I moved some things around in the previous commit and this one to better accomodate changes and reuse code.

Pretty similar to the PRCommentContext, but has some more details. And the pulls can be `None` (which causes the comparison to have 2 possible sources)