Refactor and fix task spawning in `Upload`
This replaces the manual chunking of upload processing tasks with a chunks generator.
In addition, it also makes the `is_final` logic explicit. Previously, there was a bug related to this logic that would not mark the final chunk correctly as `is_final` in case it aligned on a chunk boundary.