chore: reduce log level for step failed
When we fail to create the notification context we can't send that
notification, but that is not an error. So warning should suffice to aid in
debugging.

Typically this happens when some information is missing.
For example, a commit that is not part of a PR will fail to build the PR_COMMENT
notification context, simply because there is no PR associated with the commit.