fix: don't require BASE report for patch in comparison
This changes is related to codecov/engineering-team#2387.
Currently we require that the BASE report exists to compute the comparison, which includes patch data.
In fact we only need the HEAD report and a base commit, to have a diff from the git provider.
These changes calculate the patch coverage with those assumptions, thus we no longer need a BASE report to have patch coverage.

I also removed some timers in favor of sentry tracing