fix: type casting
I was expecting that returning a certain type from yaml validation would keep the type.
Maybe there's some to-from dict transformation happening but the type is being lost,
and then re-casting it is not giving the expected results.

So I'm improving the function to make sure that `["percentage",["absolute",1000.0]]` won't be returned anymore.

Fixes Sentry issue 5767825563