Remove double-`async_to_sync`
The `save_parallel_report_to_archive` function was recently refactored to not be `async` anymore.
However the `async_to_sync` wrapper was still doing its job, running the function in an async context on a background thread.

This has lead to an error inside that function which was also calling `async_to_sync` internally.

Fixes WORKER-P4B