perf: reuse pull and comparison
The objective here is to re-use the Pull and BundleAnalysisComparison that are loaded
as part of building contexts. Currently both the commit status and the PR comment notifications have to load these things independently.
Plus these just so happen to be the most slow parts of being a notification context.

Why flip the order in which we process the notifications?
Because the PR comment one is more restrictive, in the sense that the PULL cannot be `None`.
For the commit status it's possible that the Pull is `None`.

So by flipping them we can re-use the PR notification context "safely",
whereas the other way around we might run intro issues.