fix: UI and status report diff patch numbers
related: https://github.com/codecov/internal-issues/issues/721
related: https://github.com/codecov/engineering-team/issues/2387

We have a problem that the UI and the statuses are reporting different numbers for patch.
Part of this issue is that the `NotifyTask` (calculates patch for status) has access to the `Pull` while the `ComputeComparison` (calculates for the UI) doesn't.

This is a problem because the base for patch coverage might be different, so numbers would be different (see [here](https://github.com/codecov/worker/blob/9078de24b4dbed1bb64250b2b3864336954f310c/tasks/compute_comparison.py#L334) that we assume no action is needed cause there's nothing we can do).

So the 1st change here is to have the `NotifyTask` calculate and save patch coverage.
The second change is to cache the patch calculation so we don't compute it more than once.