Further optimize PathFixer/Tree
Profiling still reveals that building the tree is the slowest operation of processing a coverage report.
The PathFixer is very hot code, as it scales with the size of the monorepo.

This optimization replaces `dataclasses` with a simple constructor,
and replaces the `setdefault` call with manual code.

This is in the hope of making `Node` construction quicker, and avoid creating them if not needed.
This should hopefully speed up this code a bit, and lower pressure on the GC.