fix: WORKER-P8R
For the `commit_status` context it's possible that the `pull` is `None` (which is not the case for `comment`). I forgot to include that check.

This really shows that:
(1) the builders pattern is not very good, because of code duplication, and
(2) we need better tests

But for now this only fixes issue worker-p8r