Add vscode and direnv configs
This adds config files for `direnv` / `pyenv`.
With a system-wide install of `direnv`, this will automatically install and activate a `.venv` locally.

The vscode settings brings with it some recommended extensions, and default configuration to auto-fix/format on-save using `ruff`.
It also recommends the `mypy` extension, which gives you proper underlines for type errors, which should hopefully help with fixing these and improving typing going forward.