Remove usage of the `report_builder`
I think the intention of the `report_builder` setting was to build a `Report` from various database tables,
instead of from the `report_json` blob.

However, it is unclear if that is actually faster, as some of the involved tables are slow to query, and the `files_array` field was doing a storage request anyway.

So the code was still doing two storage requests (`chunks` and `files_array`) plus various SELECTs, instead of doing just two storage requests (`report_json` and `chunks`).