Make session.id == upload.id
Instead of trying very hard to allocate contiguous `session.id`s, just use the `upload.id` as the `session.id`.

The `upload.id` is guaranteed to be unique, so there should be no collisions.
It is also very large, so there should not be any collisions with existing sessions from merged/carry-forwarded reports.