Account for diff, changes and totals being legitimately `None`
The `ComparisonProxy` is loading certain data "on-demand" in case the local field is `None`.
This however means that the "on-demand" operation, like fetching diffs from the git provider, is done each and every time the result is legitimately `None`.

This is now using a `NOT_RESOLVED` sentinel object to mark whether results need to be fetched, which can then be `None` without triggering a re-fetch.