Keep track of in-progress Uploads
This introduces a new component that is responsible for keeping track of the processing status of individual uploads.

The goal here is that this state tracking can make batched Report merging and debounced notifications a lot more intentional.

The state tracking is achieved by keeping sets of `upload_id`s in Redis, one for in-progress (currently being parsed) Uploads, another for processed (and waiting to be merged) Uploads.