chore: update shared (get_PR changed to use graphQL)
This commit changes 2 things:
1. Updates shared to include https://github.com/codecov/shared/commit/6c0c2d68a0ab914242a8a03bedfcedf7a687588f
the commit in which we replaced calls to get PR info from github to use graphQL.
This breaks many tests that use cassettes, so we have to fix those.

2. Looking at tracing data in the `fetch_and_update_pull_request_information` function we make 3 requests to the DB:
the insert-update, query the recently inserted object, and refresh it.
In particular the refresh seems unecessary given that we _just_ pulled that info from the DB.

The goal then is to remove the refresh, but I didn't like the insert-with-update and changed the code to be more
a get_or_create style that doesn't access the table directly.