chore: fix (some) type errors in BA notify
This was not the main goal I had, but these errors were really distracting,
so fixing some type errors in BA notifiers.

One side-effect is that when loading the user_config we are now expecting the comment
config to be `bool | dict` (not `None`). This is in fact the case because we have default
config in place, but the tests didn't know that. Now they do.

I might have also caught a bug in the user activation in terms of the func args.