chore: improve (very little) some types
The important change here is making the `FullCommit` attributes nullable.
We have seen evidence in Sentry issues (although you will just have to take my word for it)
that these things can be `None`.
Actually even in the code when creating `FilteredComparison` you can see the `Report` might be
`None`.

In any case, after making this change I adjusted some types here and there,
and make explicit "return None" when the thing we depend on is not available.