Add metrics for intermediate Report size
Emits metrics for uncompressed and `zstd` compressed intermediate Reports.

Saving these to GCS is slow and might also be expensive. Saving them to a more transient storage such as redis should be a lot faster and potentially cheaper as well. However, such intermediate storage might be a lot more size constraint, so we want to figure out what we are dealing with.

In addition, this also cleans up a bunch of leftover code after fully rolling out parallel processing.