Do not clear the `LogContext` on task finish
Previously, this would unset the `LogContext` back to an empty version at the end of the task `run` hook.
However this would also unset all the previously set Sentry tags. As the Sentry Celery integration automatically wraps the `run` function, the Sentry tags have to persist past the end of said function.

This will now only set an initialized `LogContext` when entering `run`, and keep that around until the next task overwrites it. That way, we should also get an initialized context for any async code that runs outside of the `run` scope.