Only update the `Upload` once in `UploadFinisher`
This avoids updating the `Upload` state in the first `UploadProcessor` job, deferring that work to the final `UploadFinisher`, which was doing the same anyway.

This way, it is also possible to run the `UPDATE`, as well as all the `INSERT`s for either `UploadError` or `UploadLevelTotals` (which we should start removing) in bulk.