feat: use polars instead of arrow
this commit switches the cache test rollups task to using polars instead
of pyarrow. It seems polars uses pyarrow internally anyways but we want
to use polars in the api for additional filtering and ordering of the
results we calculate and store in the cache task, so for parity with the
api we use polars here as well.