Flush pulls to reduce chance of id collision
Flushing the pull before creating the commit seems to reduce the chance of
id collision. This is probably because commit _tries_ to create a pull.
Flushing the pull logs the pull object in the DB, which disables the
CommitFactory from also creating the pull object. This is my best understanding.