feat: create cache rollup cron task
this is a task that runs nightly and queues up cache test rollups tasks
based on the content of the last_rollup_date table.

basically if a rollup exists that was valid yesterday, we want to
recalculate the aggregates for the next day. Eventually the branch's
rollups become so out of date that we don't care about recalculating
the aggregates anymore