Optimize Report merging
This change tries to optimize the slow `Report.merge` operation.

When merging report `B` into an existing report `A`, we can just use `B` in case `A` is empty.
Also, the speed of merging depends on the size of the right hand size Report (`B` in this example).
By swapping the reports, we can make sure to always merge the smaller Report into the larger one.