Optimize repeated path-fixing and calls to `supports_labels`
Calls to the PathFixer are very hot in report processing.
I have the suspicion that some formats might call the path fixer multiple times for the same file. So it might be beneficial to go through a small cache in that case.

Previously, each call to `create_coverage_line` would call `supports_labels` which goes through all of the flags in the yaml config.
As this result never changes, this will now cache the result for the whole processor.