Fix notifiers assuming a `Report` exists
The various Notifier classes assume that a `Report` always exists, which was not the case previously, and they would fail in various ways due to `report` being `None`.

Instead of actually fixing all of the usages deep within the notifier code, this just makes sure that an (empty) `Report` indeed exists.

This is suboptimal, as we should really fix the wrong assumptions in the code. However doing so would be quite some work, and this seems like a quicker quickfix.