Enable timeseries for local testing environment
All the `Skipping timeseries migration` logs are due to the fact that a
timeseries database is configured (by default) but it's not enabled,
which doesn't trigger the migration and hence the warning.
We _could_ probably also remove the warning, as it's kind of noisy.
But this is quicker, because the warning is in `shared`.