fix(ta): handle random task retry better
the previous fix for this was not enough. what's currently happening is
that we're processing the files, marking the upload as processed then we
crash with none of the parsed testruns being persisted to the database

we need to make sure that we persist the processed tests to the database
before marking an upload as processed and then we are safe to rewrite it
as readable.

so the order for a successful parsing is:
process it -> write testruns to database -> upload.state = process -> db commit -> safe to rewrite

a failed parsing would be:
process it -> upload.state = failed -> db commit -> safe to rewrite

we never want the rewrite or the upload state to be updated before
we persist whatever processed testruns we need to persist to the
database