Handle Test/Flag inserts in the DB
This eagerly inserts all the test / flags combinations, and uses a `on_conflict_do_nothing` to avoid duplicates.

Previously, this would resolve/fetch *all* the tests that have *any* repo flag, and deduplicate those on the python side.

That code was very slow, as the underlying query was extremely slow for unknown reasons. I also believe the code was potentially buggy, as it queried for *all* the tests with repo flags, not filtering for the specific flags we were interested in. Though it was probably fine, as the flags are part of the deterministic `test_id`.

In addition, this queries for the `flag_id` directly, turning it into a `set`, instead of using a `map` and a lookup for each test.