Make `TestRollups` `SELECT FOR UPDATE` explicit
The previous usage of django `update_or_create` was implicitly creating a transaction and using a `SELECT FOR UPDATE`.

We do the same logic now explicitly, using one long-running transaction and `get_or_create`.
The new long running transaction and `FOR UPDATE NOWAIT` can fully replace the Redis-based lock in a followup.