fix(flakes): process uploads only once
we want to process test instances related to an upload only once. If an
upload has already been taken into account we don't want to take them
into account again.

we are still only processing test instances that are related to code
that is on the main branch since this task is only being run when:
- we receive a sync pulls task for the PR associated with this commit
  (merge existing test instances)
- we process test results for this commit that come from the main branch
  (we receive new relevant test instances)