Resolve notifier `pull` on demand
A recent change to load the `repo_provider` and `pull` ahead of time regressed the happy path of test results when no test failures are present.
In that case, no notification has to be performed, so no `repo_provider` or `pull` is needed.

Additionally to fixing this regression, this also fixes the case when the `pull` is legitimately `None`, but the on-demand loading was trying to fetch it once again.