Remove the Impact Analysis Deprecation from footer
This was duplicated because it's also added at the end of the message.
So, removing this from the footer to ensure that the message is on more comments.
It is being hidden by a feature flag that checks for eligible repos,
so the impact of this change should be fairly low.