Implement storage- and relationship-aware cleanup
This implements a generic mechanism to resolve the model relationship graph based on Django metadata.
That graph is then sorted in topological order, and all models are being deleted.
Special manual deletion code is written for all the models that have some files in storage, in order to make sure that files are also being cleaned up properly.