Further simplify the DELETE queries
For the case where we were deleting models with files in chunk, the actual `DELETE` query would be chained to the existing input queryset.
That way, the resulting query would have `WHERE original_filter AND id IN (original_filter...)`.
Now only the `IN` part should remain.

Also adds a bunch of tracing annotations to better see each phase, in particular these file deletions.