Generate separate queries instead of `OR` filters
In todays episode of "the DB is slow for unexplainable reasons", we found out
that querying `compare_commitcomparison` by **both** `base_commit_id` **OR** `compare_commit_id` is horribly slow.
It is using a `Seq Scan` in that case for reasons unknown.

Splitting that `OR` filter up into two separate queries makes the DB use the appropriate indices, and the queries become instant.
We can do this as the `OR` filter pretty much takes the union of both queries, which is equivalent of running both queries separately.

God only knows why, but here we are.