Delete chunked rows based on PK
Previously, I was SELECTing the first N records, deleting the associated files, and then running the DELETE query using a subquery selecting those first N records again.

The primary reason was to avoid having to pass those primary keys through python, which generates suboptimal and "unique" (for N records) queries.

This approach does not work well when running the same function concurrently multiple times, as the DELETE is not idempotent regarding the previous SELECT.

Now I did change this to actually do the pk roundtrip through python, which makes the DELETE idempotent and resilient against running this in parallel.
We might end up in a situation where parallel jobs try to delete the same records, but that is now an idempotent operation and does not cause any problems.