Add `ProfilingUpload` to the `regular_cleanup` job
Thi `regular_cleanup` cron task is a scheduled task that is used to clean up data that is not needed anymore.

In particular, it will delete all the storage associated with certain models, as opposed to just dropping the table.

As we have deprecated the "profiling"/"impact analysis" feature, and are about to drop the `ProfilingUpload` model completely, lets run it through this cleanup task to fully wipe all the uploaded data as well.