fix: process flakes redis usage
Previously, I made the decision to use a redis lock to exclude
multiple process_flakes tasks from running for the same repo at the
same time, so that we would minimize database concurrency issues.

The idea was that I would set this redis key that represents "more
flakes need to be processed for this repo" and that tasks that failed
to acquire the lock could rely on the currently running task to handle
their work so they wouldn't have to block. I never actually
implemented the part where the currently running task takes the work
of tasks that ceded their work to it.

This commit implements that by replacing the value of the current
redis key with a list of commits that need to be processed, this way
the currently running task actually has access to the commits that
need to be processed.

Since it's possible there are old versions of the ta finisher running,
our new process flakes task needs to handle both versions of the redis
scheme.

If an old process flakes task receives a task queued up by the new
task it will do nothing, but the next time a new process flakes task
runs for that repo, it will process the flakes for that commit.