Fix `report_lock` TTL
It might be possible that `hard_time_limit` is `0`, in which case we would define a `0` TTL for the redis lock.

The problem here is that we have observed some data races related to this lock apparently not working as intended. The redis key / lock TTL-ing might be one cause of this.