Remove `get_changes_using_rust` usage
As the `rust_report` is hard-enabled because `get_impacted_files` depends on it, we are also unconditionally running both implementations of `get_changes`.

But those two implementations differ in some details, and we are spamming logs with a ton of differences that noone ever looks ate to eventually move the implementation one way or the other.

So lets just remove this comparison code for a tiny speedup.