Do not consider method-lines as relevant in cobertura
The etree API is a bit weird, because the `iter()` method will iterate over matching elements in the whole subtree, instead of only over child elements.
Which we actually use extensively, as we are not fully traversing the xml structure of the various formats, but just iterate over the relevant elements, regardless of their nesting.

However for cobertura this might not be the right thing to do. As a `class` can have `lines>line`, but also a `class>methods>method` can have `lines>line` as well.
We are only interested in the directly declared `lines>line`, not the ones nested within `methods` however.

---

Along with fixing this for the cobertura processor, this also changes some tests to use the proper class. We use the `lxml` library in production, but a ton of tests were using the `xml` library, which has a mostly compatible API, but differs in some small details. For example, `lxml` has the `iterchildren` method, while `xml` does not.

Along with changing to the proper `lxml` library, some tests also changed their inputs to `bytes`, as the `lxml` library is otherwise throwing an error when trying to parse an xml declaration that defines an `encoding` from a python `str`, which is utf-8 by definition.