Move debugging test
This now defines two skipped-by-default tests that are supposed to be manually invoked with some downloaded production artifacts to more easily be able to reproduce customer issues.

There is a test to run the raw upload processing, as well as a test that just loads a stored report into memory so it can be inspected further.