fix: remove process flakes commit_sha arg
this is the final step in moving the process flakes task to using the
new method of fetching commit shas from redis. Since all TA finisher
tasks are writing to redis using the new method, the process flakes
task doesn't need to read the old method anymore, so we can remove
that.