Add regular `Upload` cleanup
This adds logic clean up `Upload`s older than 120 days currently.

The logic to do this is carefully crafted to spread out the deletion into smaller chunks, as the production `Upload` table is one of the largest we have, and bulk-deletion of all uploads will probably cause other DB issues.